refactor: split export module
